OFA/requirements.txt
aml==0.0.0
apex==0.9.10dev
azureml==0.2.7
bitarray==2.6.0
boto3==1.24.62
botocore==1.27.62
chartio==6.0.1
cython==0.29.32
deepspeed==0.7.2
editdistance==0.6.0
fairscale==0.4.8
faiss==1.5.3
fastapi==0.81.0
fastBPE==0.1.0
fasttext==0.9.2
flashlight==0.1.1
g2p_en==2.1.0
gossip==2.4.0
h5py==3.7.0
indicnlp==0.0.1
inflect==6.0.0
iopath==0.1.10
joblib==1.1.0
kaldi_io==0.9.4
kenlm==0.0.0
librosa==0.9.2
matplotlib==3.5.3
nltk==3.7
npy_append_array==0.9.13
numpy==1.23.1
omegaconf==2.2.3
pandas==1.4.3
phonemizer==3.2.1
Pillow==9.2.0
progressbar33==2.4
psutil==5.9.1
pyarrow==9.0.0
pycocoevalcap==1.2
pycocotools==2.0.4
pythainlp==3.0.8
pyworld==0.3.0
PyYAML==6.0
regex==2022.8.17
requests==2.28.1
sacrebleu==2.2.0
sacremoses==0.0.53
scikit_learn==1.1.2
scipy==1.9.0
sentencepiece==0.1.97
setuptools==58.1.0
simuleval==1.0.2
six==1.16.0
soundfile==0.10.3.post1
spacy==3.4.1
submitit==1.4.5
subword_nmt==0.3.8
tabulate==0.8.10
tensorboardX==2.5.1
tgt==1.4.4
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.64.0
transformers==4.21.2
Unidecode==1.3.4
wandb==0.13.2
webrtcvad==2.0.10
wget==3.2
